requirements.txt
Flask==3.1.0
SQLAlchemy==2.0.38
psycopg[binary]==3.2.3
python-dotenv==1.1.1
google-generativeai==0.8.5
gunicorn==23.0.0
psycopg2
psycopg2-binary
flask-cors